humanizer_api/requirements.txt
# Core dependencies for all APIs
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.0.3
python-dotenv==1.0.0
python-multipart==0.0.6

# HTTP client
httpx==0.25.2
aiohttp==3.9.0

# Rich console output
rich==13.7.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
chromadb==0.4.18

# Vector search and embeddings
sentence-transformers==2.2.2
numpy>=1.24.0
scipy>=1.10.0

# Text processing
python-markdown==3.5.1
beautifulsoup4==4.12.2
html2text==2020.1.16

# LLM integrations
openai==1.3.0
anthropic==0.7.0
ollama==0.1.7

# Async support
asyncio-mqtt==0.13.0
asyncpg==0.29.0

# Utilities
click==8.1.7
rich==13.7.0
python-dateutil==2.8.2
pyyaml==6.0.1
validators==0.22.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.0

# Memory management
chromadb-memory>=0.1.0
